feat(#27): commentCreate -> commentService, commentModel 분리
